refactor: alter setup to override fixture.beforeEach with test.before in e2e login test
