+++ test/helpers/set-fixture-hooks.js
@@ -1,0 +1,7 @@
+import { startServer, setup, teardown } from './setup'
+
+export default chosenFixture =>
+  chosenFixture
+    .before(startServer)
+    .beforeEach(setup)
+    .afterEach(teardown)
